v2.7 Toegevoegd: - binary file met daarin de stereo binaural impulse responses. - een class die de binauraldata bevat en aan de convolutieclass toerijkt. - een convolutieclass met daarin:     -  binaural panning dmv van tijdsgebaseerde convolutie     -  interpolatie tussen twee IRs om clicks te voorkomen. - binDelay gebruikt niet meer de DC bin. - refactorred blockDelay - binaural panSlider die de snelheid van de binaural panner bepaald.
 TODO list is uitgesteld. Ik moet daar later weer mee aan de slag, en dat moet ik ook alle oude todo's langslopen.
Voor nu focussen op het stabiel en werkend krijgen van het convolutie algoritme.

+++ FFTBinDelay/Source/overlapFFT.h
@@ -29,7 +29,7 @@ class OverlapFFT {
 		void pushFFTDataIntoOutputDelayBuffer(int startIndex, int endIndex); // [4]
 		
 		// [5]
-		double getOutputData();
+		float getOutputData();
 		// [6, 7]
 		void adjustMemoryPointersAndClearOutputMemory();
